support-resistance-breakout/src/requirements.txt
basebot22-basebot-JustinGuese
numpy
pandas
tqdm